event cache: move store trait and memory impl to a new `store` file
